+++ CHANGELOG.md
@@ -9,14 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- **\[Python/C++\]** Add `GetPage(VoxelKey key)` to `Writer`
 - **\[Python/C++\]** Add `PageExists(VoxelKey key)` to `Writer`
 - **\[Python/C++\]** Add `GetParentAtDepth(int32_t depth)` to `VoxelKey`
+- **\[Python/C++\]** Add `GetPageList()` to `Reader`
+- **\[Python/C++\]** Add `page` member to `Node`
+- **\[Python/C++\]** Add `ChangeNodePage` to `Writer`
 
 ### Changed
 
 - **\[Python/C++\]** Write WKT as EVLR instead of VLR to align with laspy
 - **\[Python/C++\]** Remove the error when scan angle is not within LAS specs [-30000,30000]
+- **\[Python/C++\]** !!!Potential Breakage!!! Replaced `Page page` argument with optional `VoxelKey page_key` in `Writer`'s `AddNode` and `AddNodeCompressed`. If no `page_key` is provided the node is added to the root page, if a `page_key` is provided then the node is added to that page. If the requested page doesn't exist, it is created using the nearest existing hierarchical parent as page parent.
 
 ## [2.0.0] - 2021-10-28